requirements.txt
tensorflow
tensorflow_hub
matplotlib
pillow
numpy
pytesseract
gTTS
